RF: move create_test_dataset outside of tests/
otherwise requires nose to run it... detected after trying to use datalad
installed under singularity container with no recommends installed
